Split class into two.
